back/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
exifread
hachoir
python-dotenv
